refactor(format): extract markdown git renderer
